pyproject.toml
[project]
name = "pymetal"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "einops>=0.8.1",
    "numpy>=2.3.1",
    "pyobjc>=11.1",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
]
